requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
ollama==0.1.7
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
schedule==1.2.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
asyncio==3.4.3
rich==13.7.0
typer==0.9.0
